+++ Code/gui/windows/admin/__init__.py
@@ -1,0 +1,1 @@
+from .main import admin_main_window
