+++ Content/Client/loc/rus/admin.loc
@@ -1,9 +1,12 @@
+# For main
+admin_control_menu = Админ панель
+user_access_control_lable = Управление правами
+manage_server_settings_lable = Управление настройками сервера
+manage_users_lable = Управление пользователями
+
+# For access control
 full_access_warning = У этого пользователя полный доступ. Вы не можете поменять права доступа.
-
 user_control = Управление правами пользователей
-
 users_control_logins = Пользователи:
-
 user_control_access_control = Права доступа
-
 user_control_not_load_user = Пользователь не выбран...
