+++ Content/Client/loc/rus/start_app.loc
@@ -13,4 +13,7 @@ connect_port_lable = Порт
 login_button = Авторизоваться
 register_button = Зарегестрироваться
 
+timeout_error = Время ограничения подключении вышло. Перепроверьте логин и пароль или адресс подключения
 on_connect_error_msg = Ошибка при подключении к серверу. Проверьте пароль/логин или адресс подключении
+
+cur_login = Login: {login}
