+++ requirements.txt
@@ -1,6 +1,6 @@
 dearpygui
 dearpygui-async
-DMBotNetwork==0.2.5
+DMBotNetwork==0.2.6
 DMBotTools
 msgpack
 Pillow
